refactor(util): Moved util.js in to the src folder
We want to precompile all js files with babel and for simplicity we
should keep all source files in the same folder.

The test are using `util` but I can't see any real problems with having
the tests importing a file inside the src folder
