+++ src/cli/parsers/git-cz.js
@@ -1,6 +1,6 @@
 import minimist from 'minimist';
 
-import { isString, isArray } from '../../../common/util';
+import { isString, isArray } from '../../common/util';
 
 export {
   parse
