+++ src/cli/strategies/git-cz.js
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import sh from 'shelljs';
 import inquirer from 'inquirer';
-import {getParsedPackageJsonFromPath} from '../../../common/util';
+import {getParsedPackageJsonFromPath} from '../../common/util';
 import {gitCz as gitCzParser, commitizen as commitizenParser} from '../parsers';
 import {commit, staging, adapter} from '../../commitizen';
 import {addPath} from '../../git';
